+++ stdlib/Random/src/Random.jl
@@ -310,7 +310,8 @@ Pick a random element or array of random elements from the set of values specifi
 * a string (considered as a collection of characters), or
 * a type: the set of values to pick from is then equivalent to `typemin(S):typemax(S)` for
   integers (this is not applicable to [`BigInt`](@ref)), to ``[0, 1)`` for floating
-  point numbers and to ``[0, 1)+i[0, 1)`` for complex floating point numbers;
+  point numbers and to ``[0, 1)+i[0, 1)`` for complex floating point numbers; `Tuple` types
+  are sampled component-wise.
 
 `S` defaults to [`Float64`](@ref).
 When only one argument is passed besides the optional `rng` and is a `Tuple`, it is interpreted
@@ -319,6 +320,9 @@ as a collection of values (`S`) and not as `dims`.
 
 !!! compat "Julia 1.1"
     Support for `S` as a tuple requires at least Julia 1.1.
+
+!!! compat "Julia 1.6"
+    Support for `S` as a tuple type requires at least Julia 1.6.
 
 # Examples
 ```julia-repl